Training the model
The model was first outlined, then optimizers and loss functions were set. Finally the model was trained with an early stopping.

While outlining the model, the input was first flatten into a vector. Then 2 hidden layers were added with hidden layer size of 50 and relu as the activation function. The last layer had 10 output size and softmax as the activation function. As the activation function of the output layer must transform the values into probabilities in this classification problem, we use the activation function softmax at the last.

We choose the optimizer adam as it includes learning rate schedules and momentum, and the loss function as sparse_categorical_crossentropy to ensure that our integer targets are one hot encoded.

When training the model, we set verbose =2 to observe the loss and the accuracy. The early stopping mechanism was also set to avoid overfitting.